requirements.txt
flask==2.2.3
openai==0.27.4
PyYAML==6.0